requirements.txt
# Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
rembg==2.0.50
numpy==1.24.3

# Optional: For advanced AI-based processing
# mediapipe==0.10.7

# Development Dependencies
python-decouple==3.8

# Production Dependencies (uncomment for production)
# gunicorn==21.2.0
# whitenoise==6.6.0
# psycopg2-binary==2.9.9  # For PostgreSQL
